CBP - Stock Transactions - Bank Money Destock - Implementation - QA Fix GetManager #2920
